Crate burn_compute

Burn compute crate that helps creating high performance async backends.

Modules

channel

Compute channel module.

client

Compute client module.

memory_management

Memory management module.

server

Compute server module.

storage

Compute Storage module.

tune

Autotune module

Macros

memory_id_type

Create a new memory ID type.

storage_id_type

Create a new storage ID type.

Structs

Compute

The compute type has the responsibility to retrieve the correct compute client based on the given device.