Module burn_compute::memory_management

Memory management module.

Structs

ChunkId

Memory ID.

SimpleMemoryManagement

Reserves and keeps track of chunks of memory in the storage, and slices upon these chunks.

SliceId

Memory ID.

Enums

DeallocStrategy

The strategy defines the frequency at which deallocation of unused memory chunks should occur.

SimpleHandle

The SimpleHandle is a memory handle, referring to either a chunk or a slice.

SliceStrategy

The strategy defines when to reuse chunk with slices.

Traits

MemoryHandle

The MemoryHandle trait is an abstract way to refer to some memory segment. It should not contain actual references to data.

MemoryManagement

The MemoryManagement trait encapsulates strategies for (de)allocating memory. It is bound to the ComputeStorage trait, which does the actual (de)allocations.