Trait burn_compute::memory_management::MemoryManagement

pub trait MemoryManagement<Storage: ComputeStorage>: Send + Debug {
    type Handle: MemoryHandle;

    // Required methods
    fn get(&mut self, handle: &Self::Handle) -> Storage::Resource;
    fn reserve(&mut self, size: usize) -> Self::Handle;
    fn alloc(&mut self, size: usize) -> Self::Handle;
    fn dealloc(&mut self, handle: &Self::Handle);
    fn storage(&mut self) -> &mut Storage;
}

The MemoryManagement trait encapsulates strategies for (de)allocating memory. It is bound to the ComputeStorage trait, which does the actual (de)allocations.

The MemoryManagement can only reserve memory space or get the resource located at a space. Modification of the resource data should be done directly on the resource.

Required Associated Types

type Handle: MemoryHandle

The associated type Handle must implement MemoryHandle

Required Methods

fn get(&mut self, handle: &Self::Handle) -> Storage::Resource

Returns the resource from the storage at the specified handle

fn reserve(&mut self, size: usize) -> Self::Handle

Finds a spot in memory for a resource with the given size in bytes, and returns a handle to it

fn alloc(&mut self, size: usize) -> Self::Handle

Bypass the memory allocation algorithm to allocate data directly.

Notes

Can be useful for servers that want specific control over memory.

fn dealloc(&mut self, handle: &Self::Handle)

Bypass the memory allocation algorithm to deallocate data directly.

Notes

Can be useful for servers that want specific control over memory.

fn storage(&mut self) -> &mut Storage

Fetch the storage used by the memory manager.

Notes

The storage should probably not be used for allocations since the handles won’t be compatible with the ones provided by the current trait. Prefer using the alloc and dealloc functions.

This is useful if you need to time the deallocations based on async computation, or to change the mode of storage for different reasons.