Trait burn_compute::server::ComputeServer

pub trait ComputeServer: Send + Debugwhere
    Self: Sized,{
    type Kernel: Send;
    type Storage: ComputeStorage;
    type MemoryManagement: MemoryManagement<Self::Storage>;
    type AutotuneKey: AutotuneKey;

    // Required methods
    fn read(&mut self, handle: &Handle<Self>) -> Reader<Vec<u8>>;
    fn create(&mut self, data: &[u8]) -> Handle<Self>;
    fn empty(&mut self, size: usize) -> Handle<Self>;
    fn execute(&mut self, kernel: Self::Kernel, handles: &[&Handle<Self>]);
    fn sync(&mut self);
}

The compute server is responsible for handling resources and computations over resources.

Everything in the server is mutable, therefore it should be solely accessed through the compute channel for thread safety.

Required Associated Types

type Kernel: Send

The kernel type defines the computation algorithms.

type Storage: ComputeStorage

The storage type defines how data is stored and accessed.

type MemoryManagement: MemoryManagement<Self::Storage>

The memory management type defines strategies for allocation in the storage type.

type AutotuneKey: AutotuneKey

The key used to cache operations used on specific inputs in autotune

Required Methods

fn read(&mut self, handle: &Handle<Self>) -> Reader<Vec<u8>>

Given a handle, returns the owned resource as bytes.

fn create(&mut self, data: &[u8]) -> Handle<Self>

Given a resource as bytes, stores it and returns the memory handle.

fn empty(&mut self, size: usize) -> Handle<Self>

Reserves size bytes in the storage, and returns a handle over them.

fn execute(&mut self, kernel: Self::Kernel, handles: &[&Handle<Self>])

Executes the kernel over the given memory handles.

Kernels have mutable access to every resource they are given and are responsible of determining which should be read or written.

fn sync(&mut self)

Wait for the completion of every task in the server.

Object Safety

This trait is not object safe.